feat: Prefer fetching the environment from O11Y_ENV
This is something we have already in hats, and makes sense more
generally. We previously used RAILS_ENV as the source of the
environment, but RAILS_ENV has a limited set of values which we can't
always easily change to distinguish o11y data where we want to.

In this concrete case, we added a second staging environment for
pentesters, but wanted to tag it with a different environment on
Datadog. The same is true for the load test environment, which will use
a production RAILS_ENV.