Disable ActiveRecord spans for real this time
I tested this variant of the change locally to ensure it does the right
thing.

We can't use `enabled: false` for some reason, I suspect because it's
force-enabled by the Rails instrumentation. We also can't use a span
filter, because that filters the whole subtree, and there's no easy way
of re-parenting the mysql2 spans below. This does work though,
apparently the env var has enough precedence.